requirements.txt
requests==2.33.1
psycopg2-binary==2.9.12
